修复一些警告

+++ JNWarmTips/Calendar/Model/JNDayModel.h
@@ -7,9 +7,9 @@
 
 
 @interface JNDayModel : NSObject
-@property (nonatomic, assign) NSInteger day;
-@property (nonatomic, assign) NSInteger month;
-@property (nonatomic, assign) NSInteger year;
+@property (nonatomic, assign) int day;
+@property (nonatomic, assign) int month;
+@property (nonatomic, assign) int year;
 @property(nonatomic, assign) BOOL needShowFlag;
 @property(nonatomic, assign) BOOL isCurrentMonth;
 @property(nonatomic, assign) BOOL isToday;
