+++ JNWarmTips/ToDo/Model/JNItemModel.h
@@ -12,7 +12,7 @@
 @property (nonatomic, strong) NSString *showDate;
 @property(nonatomic, assign) long long startTime;
 @property(nonatomic, assign) long long endTime;
-@property(nonatomic, assign) NSInteger groupId;
+@property(nonatomic, assign) long long groupId;
 @property(nonatomic, assign) NSInteger categoryId;
 @property(nonatomic, assign) BOOL finished;
 @property(nonatomic, assign) BOOL notification;
